Clean up type location
