+++ packages/harmony/src/components/button/Button/types.ts
@@ -1,0 +1,33 @@
+import { ComponentPropsWithoutRef } from 'react'
+
+import type { SpecialColors } from 'foundations/color'
+
+import { BaseButtonProps } from '../BaseButton/types'
+
+export type HTMLButtonProps = ComponentPropsWithoutRef<'button'>
+
+export type ButtonVariant = 'primary' | 'secondary' | 'tertiary' | 'destructive'
+
+export type ButtonSize = 'small' | 'default' | 'large'
+
+export type ButtonProps = {
+  /**
+   * Override the color of the button, only valid for the `PRIMARY` variant
+   */
+  color?: SpecialColors
+
+  /**
+   * Override the color of the button using any hex color, only valid for the `PRIMARY` variant
+   */
+  hexColor?: `#${string}`
+
+  /**
+   * The type of the button
+   */
+  variant?: ButtonVariant
+
+  /**
+   * The button size
+   */
+  size?: ButtonSize
+} & Omit<BaseButtonProps, 'styles'>
