+++ packages/harmony/src/components/button/index.ts
@@ -1,7 +1,10 @@
 export { Button } from './Button/Button'
+export * from './Button/types'
 export { PlainButton } from './PlainButton/PlainButton'
+export * from './PlainButton/types'
 export { SocialButton } from './SocialButton/SocialButton'
 export { FollowButton } from './FollowButton/FollowButton'
+export * from './FollowButton/types'
 export { FilterButton } from './FilterButton/FilterButton'
+export * from './FilterButton/types'
 export { IconButton, IconButtonProps } from './IconButton/IconButton'
-export * from './types'
